+++ frontend/src/main.jsx
@@ -4,13 +4,16 @@ import { BrowserRouter } from 'react-router-dom'
 import App from './App.jsx'
 import './index.css'
 import { StatusProvider } from './contexts/StatusContext.jsx'
+import { ThemeProvider } from './contexts/ThemeContext.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <StatusProvider>
-      <BrowserRouter>
-        <App />
-      </BrowserRouter>
-    </StatusProvider>
+    <ThemeProvider>
+      <StatusProvider>
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
+      </StatusProvider>
+    </ThemeProvider>
   </React.StrictMode>,
 )